+++ server/app.ts
@@ -10,8 +10,8 @@ import { cloudStore, cloudStoreWebhooks } from './http/cloud-store'
 import downloadTasks from './http/downloads/download-tasks'
 import downloaders, { downloaderSelfRoute } from './http/downloads/downloaders'
 import { events } from './http/events'
-import ihost from './http/image-hosting/ihost'
-import ihostConfig from './http/image-hosting/ihost-config'
+import ihostConfig from './http/image-hosting/config'
+import ihost from './http/image-hosting/images'
 import internal from './http/internal'
 import { notifications } from './http/notifications'
 import objects from './http/objects'
@@ -22,14 +22,12 @@ import { adminAudit } from './http/site/audit'
 import { authProviders } from './http/site/auth-providers'
 import { brandingAdmin, publicBranding } from './http/site/branding'
 import emailConfig from './http/site/email-config'
+import { adminSiteInvitations, publicSiteInvitations } from './http/site/invitations'
 import { adminInviteCodes, publicInviteCodes } from './http/site/invite-codes'
-import licensing from './http/site/licensing'
-import licensingAdmin from './http/site/licensing-admin'
-import { adminSiteInvitations, publicSiteInvitations } from './http/site/site-invitations'
+import { licensing, licensingAdmin } from './http/site/licensing'
 import storages from './http/site/storages'
 import system from './http/site/system'
-import { publicTeams, teams } from './http/teams'
-import { adminTeams } from './http/teams-admin'
+import { adminTeams, publicTeams, teams } from './http/teams'
 import trash from './http/trash'
 import { users } from './http/users'
 import webdav from './http/webdav'
@@ -44,7 +42,7 @@ import type { Platform } from './platform/interface'
 import { getDeployPlatform } from './runtime-platform'
 import type { Deps } from './usecases/deps'
 import { INSTANCE_TELEMETRY_CRON, reportInstanceTelemetry } from './usecases/site/instance-telemetry'
-import { ensureSitePublicOrigin } from './usecases/site/site-public-origin'
+import { ensureSitePublicOrigin } from './usecases/site/public-origin'
 
 export function createApp(platform: Platform, auth: Auth, deps: Deps = createDeps(platform)) {
   const app = new Hono<Env>()
