+++ server/http/cloud-store.integration.test.ts
@@ -4,7 +4,7 @@ import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { ZPAN_CLOUD_URL_DEFAULT } from '../../shared/constants'
 import { PUBLIC_KEYS } from '../domain/license-keys.js'
 import { adminHeaders, authedHeaders, createTestApp, seedBusinessLicense } from '../test/setup.js'
-import { cloudGiftCardsResponseSchema, cloudPackageResponseSchema } from './cloud-store-helpers.js'
+import { cloudGiftCardsResponseSchema, cloudPackageResponseSchema } from './cloud-store/helpers.js'
 
 const REFRESH_TOKEN = 'test-refresh-token'
 const INSTANCE_STORE_PATH = '/api/stores/store-test-binding'
