+++ server/http/cloud-store/traffic-metering.ts
@@ -1,13 +1,13 @@
 import type { Context } from 'hono'
-import { ZPAN_CLOUD_URL_DEFAULT } from '../../shared/constants'
-import type { Env } from '../middleware/platform'
+import { ZPAN_CLOUD_URL_DEFAULT } from '../../../shared/constants'
+import type { Env } from '../../middleware/platform'
 import {
   type DownloadTrafficOutcome,
   type DownloadTrafficStorage,
   meterDownloadTraffic,
   reportDownloadEgress,
   type TrafficReportSource,
-} from '../usecases/cloud-traffic-metering'
+} from '../../usecases/cloud-traffic-metering'
 
 // Thin http adapters over the download-metering usecase: resolve the cloud base
 // URL from the request, call the usecase (deps passed whole), and render the
