+++ server/http/image-hosting/config.integration.test.ts
