+++ server/http/site/licensing.ts
@@ -4,20 +4,44 @@ import { Hono } from 'hono'
 import { ZPAN_CLOUD_URL_DEFAULT } from '../../../shared/constants'
 import type { BindingState } from '../../../shared/types'
 import { originFromRequestUrl } from '../../domain/site-public-origin'
+import { requireAdmin } from '../../middleware/auth'
 import type { Env } from '../../middleware/platform'
 import { syncPendingCloudTrafficReports } from '../../usecases/cloud-traffic-metering'
 import { syncPendingRemoteDownloadUsageReports } from '../../usecases/downloads/remote-download-usage'
 import { buildCloudInstanceInfo, runtimeInfo } from '../../usecases/site/instance-info'
-import { loadBindingState, normalizeHost, runLicensingRefresh } from '../../usecases/site/licensing'
-import { getSitePublicOrigin } from '../../usecases/site/site-public-origin'
+import {
+  initiatePairing,
+  loadBindingState,
+  normalizeHost,
+  pollPairing,
+  runLicensingRefresh,
+  triggerRefresh,
+  unbindLicense,
+} from '../../usecases/site/licensing'
+import { getSitePublicOrigin } from '../../usecases/site/public-origin'
+
+function getCloudBaseUrl(c: Context<Env>): string {
+  return c.get('platform').getEnv('ZPAN_CLOUD_URL') ?? ZPAN_CLOUD_URL_DEFAULT
+}
+
+async function getInstanceOrigin(c: Context<Env>): Promise<string | null> {
+  return (await getSitePublicOrigin(c.get('deps'))) ?? originFromRequestUrl(c.req.url)
+}
+
+async function requireInstanceOrigin(c: Context<Env>): Promise<string> {
+  return (await getInstanceOrigin(c)) ?? new URL(c.req.url).origin
+}
 
 async function configuredPublicHost(c: Context<Env>): Promise<string | null> {
   const origin = await getInstanceOrigin(c)
   return origin ? new URL(origin).host : null
 }
 
-async function getInstanceOrigin(c: Context<Env>): Promise<string | null> {
-  return (await getSitePublicOrigin(c.get('deps'))) ?? originFromRequestUrl(c.req.url)
+async function getRequestHost(c: Context<Env>): Promise<string> {
+  const configured = await getSitePublicOrigin(c.get('deps'))
+  if (configured) return new URL(configured).host
+  const forwardedHost = c.req.header('x-forwarded-host') ?? c.req.header('host')
+  return normalizeHost(forwardedHost) ?? new URL(c.req.url).host
 }
 
 function cloudDashboardUrl(cloudBaseUrl: string): string {
@@ -30,9 +54,17 @@ function secretsMatch(provided: string, expected: string): boolean {
   return timingSafeEqual(enc.encode(provided), enc.encode(expected))
 }
 
-const app = new Hono<Env>()
+function isAuthorizedCronRequest(c: Context<Env>) {
+  const expectedSecret = c.get('platform').getEnv('REFRESH_CRON_SECRET')
+  const provided = c.req.query('secret') ?? ''
+  return Boolean(expectedSecret && secretsMatch(provided, expectedSecret))
+}
+
+// Public licensing surface — instance status plus cron-secret-authorized sync
+// runs. /status is anonymous; the cron endpoints authenticate via REFRESH_CRON_SECRET.
+export const licensing = new Hono<Env>()
   .get('/status', async (c) => {
-    const cloudBaseUrl = c.get('platform').getEnv('ZPAN_CLOUD_URL') ?? ZPAN_CLOUD_URL_DEFAULT
+    const cloudBaseUrl = getCloudBaseUrl(c)
     const currentHost =
       (await configuredPublicHost(c)) ??
       normalizeHost(c.req.header('x-forwarded-host') ?? c.req.header('host')) ??
@@ -51,7 +83,7 @@ const app = new Hono<Env>()
       return c.json({ error: 'Unauthorized' }, 401)
     }
 
-    const cloudBaseUrl = c.get('platform').getEnv('ZPAN_CLOUD_URL') ?? ZPAN_CLOUD_URL_DEFAULT
+    const cloudBaseUrl = getCloudBaseUrl(c)
     const origin = await getInstanceOrigin(c)
     const instance = origin
       ? await buildCloudInstanceInfo(c.get('deps'), {
@@ -69,7 +101,7 @@ const app = new Hono<Env>()
       return c.json({ error: 'Unauthorized' }, 401)
     }
 
-    const cloudBaseUrl = c.get('platform').getEnv('ZPAN_CLOUD_URL') ?? ZPAN_CLOUD_URL_DEFAULT
+    const cloudBaseUrl = getCloudBaseUrl(c)
     const [traffic, remoteDownload] = await Promise.all([
       syncPendingCloudTrafficReports(c.get('deps'), { cloudBaseUrl }),
       syncPendingRemoteDownloadUsageReports(c.get('deps'), { cloudBaseUrl }),
@@ -78,10 +110,59 @@ const app = new Hono<Env>()
     return c.json({ ok: true, ...traffic, remoteDownload })
   })
 
-export default app
+// Admin licensing surface — the cloud pairing handshake, manual refresh, and
+// unbinding. Every route requires an admin principal.
+export const licensingAdmin = new Hono<Env>()
+  .use(requireAdmin)
 
-function isAuthorizedCronRequest(c: Context<Env>) {
-  const expectedSecret = c.get('platform').getEnv('REFRESH_CRON_SECRET')
-  const provided = c.req.query('secret') ?? ''
-  return Boolean(expectedSecret && secretsMatch(provided, expectedSecret))
-}
+  .post('/pairings', async (c) => {
+    const pairing = await initiatePairing(c.get('deps'), {
+      baseUrl: getCloudBaseUrl(c),
+      instanceUrl: await requireInstanceOrigin(c),
+      runtime: runtimeInfo(c.get('platform')),
+    })
+    return c.json(pairing)
+  })
+
+  .get('/pairings/:code', async (c) => {
+    const result = await pollPairing(c.get('deps'), {
+      baseUrl: getCloudBaseUrl(c),
+      code: c.req.param('code'),
+      currentHost: await getRequestHost(c),
+      userId: c.get('userId')!,
+      orgId: c.get('orgId')!,
+    })
+
+    if (!result.ok) {
+      return c.json(
+        { error: 'invalid_certificate', reason: result.reason, cloud_unbind_error: result.cloudUnbindError },
+        502,
+      )
+    }
+
+    if (result.status === 'approved') {
+      return c.json({ status: 'approved' as const, edition: result.edition, cloud_store_id: result.cloudStoreId })
+    }
+
+    return c.json({ status: result.status })
+  })
+
+  .post('/refresh-runs', async (c) => {
+    const { lastRefreshAt } = await triggerRefresh(c.get('deps'), {
+      baseUrl: getCloudBaseUrl(c),
+      instanceUrl: await requireInstanceOrigin(c),
+      runtime: runtimeInfo(c.get('platform')),
+      userId: c.get('userId')!,
+      orgId: c.get('orgId')!,
+    })
+    return c.json({ success: true, last_refresh_at: lastRefreshAt })
+  })
+
+  .delete('/binding', async (c) => {
+    const { cloudUnbindError } = await unbindLicense(c.get('deps'), {
+      baseUrl: getCloudBaseUrl(c),
+      userId: c.get('userId')!,
+      orgId: c.get('orgId')!,
+    })
+    return c.json({ deleted: true, cloud_unbind_error: cloudUnbindError })
+  })
